mediatek: add factory image for ipTIME AX3000M
This device has a dual-boot partition scheme, the stock web interface
will flash only on the inactive ubi partitions which are kernel and
rootfs, newly flashed kernel didn't know the proper rootfs partition
without the cmdline replacement hack.

The staged patchset is used to replace a kernel argument at early-
boot with a different device tree option in chosen/bootargs-replace.
the device bootloader passes the active partition identifier to the
kernel bootargs, so the kernel will able to select a proper rootfs
with the identifier it passes from.

Installation:
* Upload factory image through the stock web interface.

Revert to stock firmware:
1. Check the current boot identifier with `fw_printenv | grep boot_from`.
2. Switch to the other side of boot identifier with `fw_setenv`. (A -> B, B -> A)
3. Perform `reboot`.

Signed-off-by: Rhnn Hur <hurrhnn@icmp.kr>