mediatek: add kernel cmdline replacement hack
Add kernel command line replacement hack to mediatek. Now we can
find and replace arguments in the kernel command line by setting
different device tree option in chosen/bootargs-find and replace.

This hack is used to replace a kernel argument at early-boot for
the boards where u-boot sets bootargs, and is need to be identified and replaced.

Signed-off-by: Rhnn Hur <hurrhnn@icmp.kr>